Move BuildVoteSession and DeleteLogVoteSession into vote_session.py
