+++ bot/voting/vote.py
